chore: rename ProposalList -> ProposalsList
