+++ plugins-structure/apps/politeia/src/components/ProposalsList/EmptyList.js
