+++ plugins-structure/apps/politeia/src/components/ProposalsList/InventoryList.js
