+++ plugins-structure/apps/politeia/src/components/ProposalsList/MultipleVoteInventoryList.js
